Add working hour contract
Change-type: minor
