+++ lib/contracts/working-hours.spec.ts
@@ -1,4 +1,4 @@
-import { asTimeSlot } from './agent-settings';
+import { asTimeSlot } from './working-hours';
 
 describe('asTimeSlot', () => {
 	test('should return full list of half hour time slots', async () => {
